+++ src/com/autoStock/analysis/results/ResultsCCI.java
@@ -13,11 +13,11 @@ public class ResultsCCI extends ResultsBase {
 	
 	public Date[] arrayOfDates;
 	public double[] arrayOfCCI;
-	public float[] arrayOfPrice;
+	public double[] arrayOfPrice;
 	
 	public ResultsCCI(int length){
 		this.arrayOfDates = new Date[length];
 		this.arrayOfCCI = new double[length];
-		this.arrayOfPrice = new float[length];
+		this.arrayOfPrice = new double[length];
 	}
 }
