+++ src/com/autoStock/analysis/results/ResultsMACD.java
@@ -11,14 +11,14 @@ import java.util.Date;
  */
 public class ResultsMACD extends ResultsBase {
 	public Date[] arrayOfDates;
-	public float[] arrayOfPrice;	
+	public double[] arrayOfPrice;	
 	public double[] arrayOfMACD;
 	public double[] arrayOfMACDSignal;
 	public double[] arrayOfMACDHistogram;
 	
 	public ResultsMACD(int length){
 		this.arrayOfDates = new Date[length];
-		this.arrayOfPrice = new float[length];
+		this.arrayOfPrice = new double[length];
 		this.arrayOfMACD = new double[length];
 		this.arrayOfMACDSignal = new double[length];
 		this.arrayOfMACDHistogram = new double[length];
