+++ src/com/autoStock/analysis/results/ResultsRSI.java
@@ -13,11 +13,11 @@ public class ResultsRSI extends ResultsBase {
 	
 	public Date[] arrayOfDates;
 	public double[] arrayOfRSI;
-	public float[] arrayOfPrice;
+	public double[] arrayOfPrice;
 	
 	public ResultsRSI(int length){
 		this.arrayOfDates = new Date[length];
 		this.arrayOfRSI = new double[length];
-		this.arrayOfPrice = new float[length];
+		this.arrayOfPrice = new double[length];
 	}
 }
